requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
kiwisolver==1.4.5
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.12.0
six==1.16.0
sortedcontainers==2.4.0
gudhi==3.9.0
pandas==2.1.4
plotly==5.18.0
scikit_learn==1.4.1.post1
scipy==1.11.4
umap_learn==0.5.5
